requirements.txt
fastapi
uvicorn
pandas
openai
pydantic
requests
python-dotenv